Extract test framework provider interface with default impl
