+++ plugin/src/main/java/org/gradle/testretry/internal/executer/framework/TestFrameworkProvider.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.testretry.internal.executer.framework;
+
+import org.gradle.api.internal.tasks.testing.TestFramework;
+import org.gradle.api.internal.tasks.testing.filter.DefaultTestFilter;
+
+interface TestFrameworkProvider {
+
+    TestFramework testFrameworkFor(DefaultTestFilter failedTestsFilter);
+
+    class ProviderForCurrentGradleVersion implements TestFrameworkProvider {
+
+        private final TestFramework testFramework;
+
+        ProviderForCurrentGradleVersion(TestFramework testFramework) {
+            this.testFramework = testFramework;
+        }
+
+        @Override
+        public TestFramework testFrameworkFor(DefaultTestFilter failedTestsFilter) {
+            return testFramework.copyWithFilters(failedTestsFilter);
+        }
+    }
+
+}
